Moved Button class into root common.components package
